move DateTime to tantivy_common
